Changes Updated - 8

+++ base/views.py
@@ -1,8 +1,10 @@
 from django.shortcuts import render, redirect
+from django.http import HttpResponse
 from django.contrib import messages
 from django.contrib.auth import authenticate,login,logout
 from django.db.models import Q
 from django.contrib.auth.models import User
+from django.contrib.auth.decorators import login_required
 
 
 from .models import Room, Topic
@@ -17,10 +19,13 @@ from .forms import RoomForm
 # ]
 
 def loginPage(request):
+    if request.user.is_authenticated:
+            return redirect('home');
+
     if request.method == 'POST':
         username = request.POST.get('username');
         password = request.POST.get('password');
-    
+
         try:
             user = User.objects.get(username=username);
         except:
@@ -63,6 +68,7 @@ def room(request,pk):
     return render(request,'base/room.html',context);
 
 
+@login_required(login_url='login')
 def createRoom(request):
     form = RoomForm();
     if request.method == 'POST':
@@ -75,9 +81,13 @@ def createRoom(request):
     return render(request,'base/room_form.html',context);
 
 
+@login_required(login_url='login')
 def updateRoom(request,pk):
     room = Room.objects.get(id=pk);
     form = RoomForm(instance=room);
+
+    if request.user != room.host:
+        return HttpResponse('You are not allowed here!');
 
     if request.method == 'POST':
         form = RoomForm(request.POST, instance=room);
@@ -89,8 +99,13 @@ def updateRoom(request,pk):
     return render(request,'base/room_form.html',context);
 
 
+@login_required(login_url='login')
 def deleteRoom(request,pk):
     room = Room.objects.get(id=pk);
+
+    if request.user != room.host:
+        return HttpResponse('You are not allowed here!');
+
     if request.method == 'POST':
         room.delete();
         return redirect('home');
